Add files via upload
FinaBoost is a website that aims to generate financial management skills among people, specifically- young adults. Earning money is tough, undoubtedly, but managing money is equally important. Hence, FinaBoost was conceived to inculcate the nuances of money-management so that they turn up to be financially-responsible citizens. The website features a home page, an about page, a contact page and pages on two important topics that address the steps towards financial management and emergency management. Also, it features a calculator that helps the users to calculate the effect of compounding interest on their money. As young collegians who’re probably going to live alone for the first time, with limited money and flexible expenditures, the ability and the skill to manage finances is utmost important to be instilled and this website aims to do just that.